requirements.txt
altgraph==0.10.2
awscli==1.11.185
backports-abc==0.5
bdist-mpkg==0.5.0
botocore==1.7.43
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.7
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.11.3
django-cors-headers==2.1.0
djangorestframework==3.7.7
docutils==0.14
ez-setup==0.9
futures==3.1.1
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
macholib==1.5.1
MarkupSafe==1.0
matplotlib==1.3.1
modulegraph==0.10.4
nose==1.3.7
numpy==1.14.0
pbr==3.1.1
psycopg2==2.7.3.1
py2app==0.7.3
pyasn1==0.3.7
pyparsing==2.2.0
python-2.7.10
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.12
singledispatch==3.4.0.3
six==1.11.0
tornado==4.5.3
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
